web-wallet: Modules `$lib/stores` and `$lib/wallet` are importing each other
Resolves #2771
