+++ web-wallet/src/lib/wallet/index.js
@@ -3,6 +3,5 @@ export { default as encryptMnemonic } from "./encryptMnemonic";
 export { default as getSeedFromMnemonic } from "./getSeedFromMnemonic";
 export { default as initializeWallet } from "./initializeWallet";
 export { default as notesArrayToMap } from "./notesArrayToMap";
-export { default as nullifiersDifference } from "./nullifiersDifference";
 export { default as profileGeneratorFrom } from "./profileGeneratorFrom";
 export { default as refreshLocalStoragePasswordInfo } from "./refreshLocalStoragePasswordInfo";
